// src/server/model/iLedger.ts
/**
 * 
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { IAttributeGroupLine } from './iAttributeGroupLine';
import { IContact } from './iContact';
import { IDocument } from './iDocument';
import { ILedgerLabel } from './iLedgerLabel';
import { IOpeningBalanceTx } from './iOpeningBalanceTx';

export interface ILedger { 
    location?: string;
    type?: number;
    disabled?: boolean;
    labels?: Array<ILedgerLabel>;
    systemGenerated?: boolean;
    ledgerGroupName?: string;
    contactAddress?: string;
    openingBalance?: number;
    amountCreditLimits?: number;
    verificationDate?: Date;
    noOfCreditDays?: number;
    depreciationRate?: number;
    serviceTaxNumber?: string;
    unit?: string;
    gstin?: string;
    printname?: string;
    contact?: IContact;
    cstNumber?: string;
    documentList?: Array<IDocument>;
    ledgerGroupId?: number;
    profitShare?: number;
    panNumber?: string;
    attrGroupLine?: IAttributeGroupLine;
    openingBalTx?: IOpeningBalanceTx;
    gstStatus?: ILedger.GstStatusEnum;
    salesManName?: string;
    routeId?: number;
    hasCostCenter?: boolean;
    costCenter?: string;
    aliasName?: string;
    salesManId?: number;
    routeName?: string;
    tinNumber?: string;
    gstType?: ILedger.GstTypeEnum;
    name?: string;
    description?: string;
    modificationdate?: Date;
    impCompanyGuid?: string;
    productUserName?: string;
    productUserId?: number;
    logInfo?: string;
    importedId?: number;
    creationdate?: Date;
    id?: number;
    jacksontype?: string;
}
export namespace ILedger {
    export type GstStatusEnum = 'VALID' | 'INVALID' | 'NOT_VERIFIED';
    export const GstStatusEnum = {
        VALID: 'VALID' as GstStatusEnum,
        INVALID: 'INVALID' as GstStatusEnum,
        NOTVERIFIED: 'NOT_VERIFIED' as GstStatusEnum
    };
    export type GstTypeEnum = 'Regular' | 'Composition' | 'Unregistered';
    export const GstTypeEnum = {
        Regular: 'Regular' as GstTypeEnum,
        Composition: 'Composition' as GstTypeEnum,
        Unregistered: 'Unregistered' as GstTypeEnum
    };
}